requirements.txt
rich>=13.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0